Print thinning and culling when transplanting.
When planting seeds indoors or out, we know that some seeds may not sprout.
When seeding directly outdoors, we seed more than we need to, and thin after the
plants are big enough to remove.  When we're starting indoors, we start more
seeds, and only take the best seedlings to transplant outdoors.

Signed-off-by: Sarah Sharp <sarah@thesharps.us>