Modify input file format for seed sprouting.
The field "average days to germination" in the input csv file will now become
two fields: "minimum days to germination" and "maximum days to germination".
Most seed packets specify the days in this format, like "6-12 days for sprouts
to emergence" or "Days to germination: 7-14."

This makes it easier for the user to input the germination days, and also allows
us to output something more interesting in the sprouting calendar.  Now we can
use the maximum days to germination to calculate the last day the plants have a
chance to emerge.  After that date, the gardener knows they should replant
(possibly with a newer seed packet), or buy seedlings.

Signed-off-by: Sarah Sharp <sarah@thesharps.us>