Start of code for frost alert app.
This is a small python program I'm working on to pull temperature data
from noaa.gov and check if temperature levels will dip low enough for a
frost to form.

This will be part of a larger phone app to alert the user via txt,
email, or flashing lights that they need to cover their spring seedlings, bring in
autumn fruits or winter squash, wrap plastic around their tomatoes to
extend the harvest, or otherwise prevent frost damage.

Signed-off-by: Sarah Sharp <sarah@thesharps.us>