Rearrange the main screen to add cold plant.
Make the main screen display:

Enter your zipcode:
[ text box ]
Configure Warnings:
[ cold plant pic ] [ text box ] °F   Cold Temperature

Jamey says it makes it easier for the user to associate the cold
temperature warning with the cold plant icon.  He expected weather apps
to base their icons on what the actual weather was, not whether the user
thought it was warm or cold.

Signed-off-by: Sarah Sharp <sarah@thesharps.us>