Make main view update input boxes.
When the main view pops up with the zipcode and minimum temperature
input boxes, update the text of those boxes if the application has
those values saved.  I needed this if I call the ColdSnap activity when
the configuration button is pushed.  I'm not sure if this is exactly the
right thing to do, as this would start to make a very deep stack.

Signed-off-by: Sarah Sharp <sarah@thesharps.us>