Create actions for refresh and config buttons.
Make the refresh button update the temperatures.  This may take
awhile, and the button is "pressed" until the update completes.

Make the setup button start the ColdSnap activity to bring up the main
screen.  This may not be the best thing, since I think this can
create a very deep stack.  We probably just need to pop this activity off
the stack instead.

Signed-off-by: Sarah Sharp <sarah@thesharps.us>