Make exception handling better, fetch latlong less.
Make sure the app can handle the internet being down gracefully.  If the
functions to fetch the latitude and longitude string fail, don't try to
split the string by the commas.  If the temperature fetch fails for some
reason, show the user "Never" and "-400" for day and temperature.

That's not the greatest thing, but we can change the interface later to
display old data.  We should also show when the data was last fetched.

Instead of fetching the latitude and longitude string from the
zipcode every time we need to get the temperature, make it be fetched once
during startup, any time we try to grab the temperatures and it
hasn't been fetched, and any time the zipcode is changed by the user.

Signed-off-by: Sarah Sharp <sarah@thesharps.us>