Garduino: Water plant when soil is dry.
Water is a good conductor of electricity.  As soil becomes dry, the water
content drops, and the resistance of the soil increases.  That means that the
measured voltage drop across the soil moisture sensor will decrease as the soil
becomes dry.

When the voltage drops below 2.5V, turn on digital pin 3, which is hooked to a
water pump.

Todo: only turn on the water pump for a few seconds.  Wait at least an hour
before attempting to measure and water again.  Use a real soil moisture sensor,
calibrated for my soil, instead of twiddling a potentiometer.

Signed-off-by: Sarah Sharp <sarah@thesharps.us>