Garduino: Water for 5 sec every 30 sec at most.
Don't wait until the voltage increases to turn off the pump.  The cheapo
soil moisture sensors are much too slow to change for that.  Instead,
water the plant for 5 seconds, which is enough for small plants.

Only allow the plant to be watered once every 30 seconds.  Later this
will be a stricter limit, like twice per day, but we need it more often
for demo purposes.

Delay 100 milliseconds between voltage reads, as we overflow the serial
port with ADC conversions if there isn't a delay.  The one second delay
was just too slow for demo purposes.  May be tweaked later.

Signed-off-by: Sarah Sharp <sarah@thesharps.us>